// src/app/administration/models/account-details.ts
export interface IAccountDetails {
  accountId: string;
  createdOn: string;
  modifiedOn: string;
  countryCode: string;
  currencyCode: string;
  subscriptionId: string;
  accountType: string;
  accountStatus: string;
  balance: number;
  alertEmailAddress: string;
  alertPhoneNumber: string;
  name: string;
  shortDescription: string;
  longDescription: string;
  alertLevel: string;
  alertChannel: string;
  callBackUri: string;
  accountPackage: string;
}